Move: 프로필 관련 컴포넌트 MyPage 폴더에서 Layout으로 이동 #120
